% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{.add_group_codings}
\alias{.add_group_codings}
\title{Adds group codings for predictive df.}
\usage{
.add_group_codings(gs, newdata)
}
\arguments{
\item{gs}{group spec}

\item{newdata}{data frame}
}
\value{
data frame, with numerics where group matches, and NAs otherwise.
}
\description{
Adds group codings for predictive df.
}
\author{
Stephen Martin
}
\keyword{internal}
