% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentMelsm.R
\name{.combine_RHS}
\alias{.combine_RHS}
\title{Combines multiple formulas' RHS into one.}
\usage{
.combine_RHS(flist)
}
\arguments{
\item{flist}{}
}
\value{
RHS formula.
}
\description{
Combines multiple formulas' RHS into one.
}
\author{
Stephen R. Martin
}
\keyword{internal}
