% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.get_elapsed_time}
\alias{.get_elapsed_time}
\title{Gets elapsed time.}
\usage{
.get_elapsed_time(sOut)
}
\arguments{
\item{sOut}{lmmelsm object.}
}
\value{
Numeric.
}
\description{
Gets elapsed time.
}
\author{
Stephen R. Martin
}
\keyword{internal}
