% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentMelsm.R
\name{.get_formula_names}
\alias{.get_formula_names}
\title{Get names in formula.}
\usage{
.get_formula_names(flist, formula = TRUE)
}
\arguments{
\item{flist}{List of formulas.}

\item{formula}{Logical. Whether to return the raw RHS (TRUE) or the vars needed (FALSE).}
}
\value{
List of lists.
}
\description{
Get names in formula.
}
\author{
Stephen R. Martin
}
\keyword{internal}
