% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentMelsm.R
\name{.get_indicator_spec}
\alias{.get_indicator_spec}
\title{Get indicator spec for stan model.}
\usage{
.get_indicator_spec(mm.inds, flist)
}
\arguments{
\item{mm.inds}{Model matrix. Complete data of all non-duplicated indicators for all factors.}

\item{flist}{Formula list.}
}
\value{
List with J_f (vector; Number of Indicators for each factor) and F_ind (matrix; each row contains the columns in mm.inds for the row-th factor).
}
\description{
Get indicator spec for stan model.
}
\author{
Stephen R. Martin
}
\keyword{internal}
