% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentMelsm.R
\name{.get_LHS}
\alias{.get_LHS}
\title{Get LHS variable as string.}
\usage{
.get_LHS(formula)
}
\arguments{
\item{formula}{Formula.}
}
\value{
String.
}
\description{
Get LHS variable as string.
}
\author{
Stephen R. Martin
}
\keyword{internal}
