% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentMelsm.R
\name{.get_RHS}
\alias{.get_RHS}
\title{Get RHS terms or variables.}
\usage{
.get_RHS(formula, terms = TRUE)
}
\arguments{
\item{formula}{Formula.}

\item{terms}{Logical (Default: TRUE). Whether to return the terms (TRUE) or the variables (FALSE). E.g., "I(x^2)" vs "x"}
}
\value{
Character vector.
}
\description{
Get RHS terms or variables.
}
\author{
Stephen R. Martin
}
\keyword{internal}
