% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.list_zip}
\alias{.list_zip}
\title{Zip two lists together with function.}
\usage{
.list_zip(..., f = cbind)
}
\arguments{
\item{...}{Lists to zip.}

\item{f}{Function}
}
\value{
List.
}
\description{
Zip two lists together with function.
}
\author{
Stephen Martin
}
\keyword{internal}
