% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.magicsep}
\alias{.magicsep}
\title{Convert char vector to columns.}
\usage{
.magicsep(charvec, labs, types = NULL, sep = "MAGICSEP")
}
\arguments{
\item{charvec}{Character vector. Characters to separate into columns.}

\item{labs}{Character vector. Labels for the columns.}

\item{types}{List (Default: All characters). List (in order) of extracted types.}

\item{sep}{String (Default: "MAGICSEP").}
}
\value{
data.frame
}
\description{
Converts a character vector int columns.
}
\details{
A wrapper around strcapture.
Given "a_MAGICSEP_b", returns a vector of "a_" and "_b".
Useful for converting a multiple-parameter string into columns of parameters.
User can give column names in \code{labs}, and types of each extracted component as a list in \code{types}.
User can customize \code{sep}, but by default assumes MAGICSEP.
}
\author{
Stephen R. Martin
}
\keyword{internal}
