% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.newline}
\alias{.newline}
\title{Print newline.}
\usage{
.newline(n = 1)
}
\arguments{
\item{n}{Integer. Number of lines to print.}
}
\value{
NULL.
}
\description{
Print newline.
}
\author{
Stephen R. Martin
}
\keyword{internal}
