% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentMelsm.R
\name{.parse_formula.indicators}
\alias{.parse_formula.indicators}
\title{Compute indicator data.}
\usage{
.parse_formula.indicators(mlist, mf)
}
\arguments{
\item{mlist}{List of measurement formulas.}

\item{mf}{Data frame with complete cases.}
}
\value{
Named List.
}
\description{
Compute indicator data.
}
\author{
Stephen R. Martin
}
\keyword{internal}
