% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentMelsm.R
\name{.parse_formula.predictor}
\alias{.parse_formula.predictor}
\title{Compute predictor data.}
\usage{
.parse_formula.predictor(plist, mf, group)
}
\arguments{
\item{plist}{List of prediction formulas.}

\item{mf}{Data frame containing complete cases.}

\item{group}{Grouping data.}
}
\value{
Named list.
}
\description{
Compute predictor data.
}
\author{
Stephen R. Martin
}
\keyword{internal}
