% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentMelsm.R
\name{.parse_formula}
\alias{.parse_formula}
\title{Convert spec to stan data.}
\usage{
.parse_formula(formulaList, group, data)
}
\arguments{
\item{formulaList}{Formula or list of formulas.}

\item{group}{Group symbol.}

\item{data}{Data frame.}
}
\value{
List.
}
\description{
Convert spec to stan data.
}
\author{
Stephen R. Martin
}
\keyword{internal}
