% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.sep}
\alias{.sep}
\title{Print separator.}
\usage{
.sep(sep = "\\n-----", n = 1)
}
\arguments{
\item{sep}{String (Default = "-----"). Separator.}

\item{n}{Integer (Optional). Number of newlines to call after sep.}
}
\value{
NULL.
}
\description{
Print separator.
}
\author{
Stephen R. Martin
}
\keyword{internal}
