% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{.simulate.X}
\alias{.simulate.X}
\title{Simulate covariates without correlation.}
\usage{
.simulate.X(n, K, P, L2_pred_only)
}
\arguments{
\item{n}{Number of observations per k unit.}

\item{K}{Number of k units (Level 2 units).}

\item{P}{Number of covariates.}

\item{L2_pred_only}{Logical (Default: FALSE). If generated predictors should be level-2 only.}
}
\value{
Matrix.
}
\description{
Simulate covariates without correlation.
}
\author{
Stephen R. Martin
}
\keyword{internal}
