% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.summary_rearrange}
\alias{.summary_rearrange}
\title{Rearrange summary output.}
\usage{
.summary_rearrange(x, cols, arrange = FALSE)
}
\arguments{
\item{x}{Summary table.}

\item{cols}{Columns, in order, to place in front.}

\item{arrange}{Logical (Default: FALSE). Whether to sort rows.}
}
\value{
Data.frame.
}
\description{
Rearrange summary output.
}
\author{
Stephen R. Martin
}
\keyword{internal}
