% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentMelsm.R
\name{.which_location_scale}
\alias{.which_location_scale}
\title{Check for location-scale formulas}
\usage{
.which_location_scale(flist, reduce = TRUE)
}
\arguments{
\item{flist}{Formula list.}
}
\value{
Numeric vector (length 2); which formulas in flist correspond to location, scale.
}
\description{
Check for location-scale formulas
}
\author{
Stephen R. Martin
}
\keyword{internal}
