% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{fitted.lmmelsm}
\alias{fitted.lmmelsm}
\title{Extracted model fitted variates.}
\usage{
\method{fitted}{lmmelsm}(object, prob = 0.95, summarize = TRUE, ...)
}
\arguments{
\item{object}{lmmelsm object.}

\item{prob}{Numeric (Default: .95). Amount of probability mass contained in the credible interval.}

\item{summarize}{Logical (Default: TRUE). Whether to return posterior summaries (TRUE) or MCMC samples (FALSE).}

\item{...}{Not used.}
}
\value{
List of eta and eta_logsd. If summarize is \code{TRUE}, then the list contains formatted summary tables. If \code{FALSE}, then the list contains MCMC samples for all variables.
}
\description{
Extract model fitted variates.
}
\details{
Extracts model fitted variates. When a latent MMELSM, these are the latent score expectations and log standard deviations. When an observed MMELSM, these are the observed score expectations and log standard deviations.
}
\author{
Stephen Martin
}
