% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMMELSM-package.R
\docType{package}
\name{LMMELSM-package}
\alias{LMMELSM-package}
\alias{LMMELSM}
\title{The 'LMMELSM' package.}
\description{
Fits latent (or observed) multivariate (or univariate) mixed effects location scale models.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
