% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo.lmmelsm}
\alias{loo.lmmelsm}
\alias{loo}
\title{loo method for LMMELSM objects.}
\usage{
\method{loo}{lmmelsm}(x, type = c("observation", "group"), ...)
}
\arguments{
\item{x}{lmmelsm object.}

\item{type}{String (Default: "observation"). If "observation", then loo is leave-row-out. If "group", then loo is leave-group-out.}

\item{...}{Not used.}
}
\value{
loo object.
}
\description{
loo method for LMMELSM objects.
}
\author{
Stephen R. Martin
}
