% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{nlist}
\alias{nlist}
\title{Creates named list.}
\usage{
nlist(...)
}
\arguments{
\item{...}{Objects for list.}
}
\value{
Named List.
}
\description{
Creates named list.
}
\author{
Stephen R. Martin
}
\keyword{internal}
