% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.lmmelsm}
\alias{print.lmmelsm}
\title{Print method for lmmelsm objects.}
\usage{
\method{print}{lmmelsm}(x, ...)
}
\arguments{
\item{x}{lmmelsm object.}

\item{...}{Not used.}
}
\value{
x (Invisibly).
}
\description{
Print method for lmmelsm objects.
}
\author{
Stephen R. Martin
}
