% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.lmmelsm}
\alias{print.summary.lmmelsm}
\title{Print method for summary.lmmelsm objects.}
\usage{
\method{print}{summary.lmmelsm}(x, ...)
}
\arguments{
\item{x}{summary.lmmelsm object.}

\item{...}{Not used.}
}
\value{
x (Invisibly).
}
\description{
Print method for summary.lmmelsm objects.
}
\author{
Stephen R. Martin
}
