% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathutils.R
\name{row_multiply_list_mats}
\alias{row_multiply_list_mats}
\title{Multiply a row by a list of matrices}
\usage{
row_multiply_list_mats(r, mats)
}
\arguments{
\item{r}{Row vector.}

\item{mats}{}
}
\value{
List of row by matrix products.
}
\description{
Multiply a row by a list of matrices
}
\author{
Stephen R Martin
}
\keyword{internal}
