% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.uni.fe}
\alias{simulate.uni.fe}
\title{Simulate univariate fixed effects predictive LM-MELSM data.}
\usage{
\method{simulate}{uni.fe}(
  n,
  K,
  lambda,
  resid,
  nu,
  mu_beta = NULL,
  logsd_beta = NULL,
  mu_logsd_cor = diag(1, 2, 2),
  mu_logsd_sigma = rep(1, 2),
  L2_pred_only = FALSE,
  X_loc = NULL,
  X_sca = NULL
)
}
\arguments{
\item{n}{Integer. Number of observations per group.}

\item{K}{Integer. Number of groups.}

\item{lambda}{Vector. Loadings.}

\item{resid}{Vector. Residual SD.}

\item{nu}{Vector. Intercepts.}

\item{mu_beta}{Vector. \code{P} Predictive coefficients for latent location.}

\item{logsd_beta}{Vector. \code{Q} predictive coefficients for latent IIV.}

\item{mu_logsd_cor}{Matrix. 2x2 correlation matrix between location and scale.}

\item{mu_logsd_sigma}{Vector (Length 2). SDs for mu and logsd RE's.}

\item{L2_pred_only}{Logical (Default: FALSE). If generated predictors should be level-2 only.}

\item{X_loc}{Matrix (Optional). If supplied, will be used as the location predictors.}

\item{X_sca}{Matrix (Optional). If supplied, will be used as the scale predictors.}
}
\value{
List of data and params.
}
\description{
Simulate univariate fixed effects predictive LM-MELSM data.
}
\author{
Stephen R. Martin
}
\keyword{internal}
