% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.lmmelsm}
\alias{summary.lmmelsm}
\title{Summary method for lmmelsm objects.}
\usage{
\method{summary}{lmmelsm}(object, prob = 0.95, ...)
}
\arguments{
\item{object}{lmmelsm object.}

\item{prob}{Numeric (Default: .95). Amount of probability mass contained in the credible interval.}

\item{...}{Not used.}
}
\value{
summary.lmmelsm object. A list containing \code{meta} (metadata) and \code{summary} (summary tables).
}
\description{
Summary method for lmmelsm objects.
}
\author{
Stephen R. Martin
}
