% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMMELSM-package.R
\docType{package}
\name{LMMELSM-package}
\alias{LMMELSM-package}
\alias{LMMELSM}
\title{The 'LMMELSM' package.}
\description{
Fits latent (or observed) multivariate (or univariate) mixed effects location scale models.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/stephenSRMMartin/LMMELSM/issues}
}

}
\author{
\strong{Maintainer}: Stephen Martin \email{stephenSRMMartin@gmail.com} (\href{https://orcid.org/0000-0001-8085-2390}{ORCID})

Authors:
\itemize{
  \item Philippe Rast \email{rast.ph@gmail.com} (\href{https://orcid.org/0000-0003-3630-6629}{ORCID})
}

}
\keyword{internal}
